Conflicts:
	arch/arm64/boot/dts/qcom/sdmmagpie-camera-sensor-qrd.dtsi
	arch/arm64/boot/dts/qcom/sdmmagpie-gpu.dtsi
	arch/arm64/boot/dts/qcom/sdmmagpie.dtsi
	arch/arm64/boot/dts/qcom/sm6150-idp.dtsi
	drivers/gpu/drm/msm/dp/dp_hdcp2p2.c
	drivers/platform/msm/gsi/gsi.c
	drivers/power/supply/qcom/smb5-lib.c
	drivers/power/supply/qcom/smb5-lib.h

+++ drivers/crypto/msm/qce50.c
@@ -6013,7 +6013,7 @@ static int qce_smmu_init(struct qce_device *pce_dev)
 		dev->dma_parms = devm_kzalloc(dev,
 				sizeof(*dev->dma_parms), GFP_KERNEL);
 	dma_set_max_seg_size(dev, DMA_BIT_MASK(32));
-	dma_set_seg_boundary(dev, DMA_BIT_MASK(64));
+	dma_set_seg_boundary(dev, (unsigned long)DMA_BIT_MASK(64));
 
 	pce_dev->smmu_mapping = mapping;
 	return ret;
